+++ docs/history.md
@@ -11,6 +11,7 @@
   * Custom events
 * Fixed bugs with deposit chets
 * Fixed bug with farming
+* Added random shuffle farming area (when you have more 1 bot farming same area configuration can start for different sites)
 ## 0.1.9
 # Added in farmerjob new plants
   * Wheat + Melon + Pumpkin